+++ gatelogue-types-py/src/gatelogue_types/__about__.py
@@ -1,2 +1,2 @@
-__version__ = "2.0.2"
-__data_version__ = 8
+__version__ = "2.0.3"
+__data_version__ = 9