+++ CHANGELOG.md
@@ -4,6 +4,13 @@
 * `RailConnection`, `SeaConnection`, `BusConnection` as nodes instead of edges
 * add rail, sea and bus to client/data viewer
 
+## v2.0.3+9 (20250601)
+* various refactors inside `gatelogue-aggregator`
+
+### Data v9
+* `ref_stop`, `ref_station` removed from `BusLine`, `RailLine`, `SeaLine`
+* they have been replaced with `stops` and `stations` which are all the stops/stations the line stops at
+
 ## v2.0.2+8 (20250505)
 * `gatelogue-types` (rs): fix `local` field missing in `SeaCompany` and extra in `AirAirline`